Fix cluster test flakiness
